Unify types on both sides of the operator and fix the size_t comparison

+++ src/js_native_api_v8.cpp
@@ -2924,7 +2924,7 @@ JSVM_Status OH_JSVM_CallFunction(JSVM_Env env,
      if (UNLIKELY(env->debugFlags)) {
         if (UNLIKELY(env->debugFlags & (1 << JSVM_SCOPE_CHECK))) {
             if (argv != nullptr) {
-                for (int i = 0; i <= argc; i++) {
+                for (size_t i = 0; i <= argc; i++) {
                     ADD_VAL_TO_SCOPE_CHECK(env, argv[i]);
                 }
             }
